Factor out determineLandlordPageOptions().
